Making bulkloop work.
