+++ examples/bulkloop/descriptors.strings
